Finished implementing statistics
Implemented the remaining two sets of statistics:
- user total (given a user ID)
- users’ average

Moreover, this commit adds code to generate and store a deviceID to the
user’s device, which is used to store his total number of clicks in the
database. The storage is persistent, and is only deleted when the app
is removed (apparently).